+++ RELEASE_NOTES.md
@@ -3,6 +3,8 @@
 1.1.1
 -----
 * fixes sample acls updates, adds option to remove acls
+* add 'propagate_links' option to import_samples method. When this option is turned on, the import
+method will propagate data links from previous samples
 
 1.0.1
 -----